fix: разделение по компонентам инфо и лайка
